Move posts sagas to a single file
