+++ client/src/components/posts/PostShow.js
@@ -7,7 +7,6 @@ import { loadSinglePost } from "actions";
 class PostShow extends Component {
   componentDidMount() {
     const { id } = this.props.match.params;
-    console.log(id);
     this.props.loadSinglePost(id);
   }
